a wandering clickles
exactly as the video put on youtube shows, this is just clickles and he'll walk in random directions! (to anyone who wants to try; try and get clickles from one side of the map to the other! he starts ins the top left "home" and if you position your player properly you'll be able to let him out!)

Signed-off-by: tydinocreations <tydino909@gmail.com>